pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "rdoc-fmri-quality-control"
version = "0.1.0"
description = "Temporal-SD based QC for rdoc_fmri scans"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "nibabel",
  "matplotlib",
  "pyyaml",
  "flywheel-sdk",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
